Rename type file home.ts to study.ts
