+++ types/study.ts
